Enable arm64 unwinding with frame pointers
(This is still early on, not very well tested so far).

We want to properly support arm64 and we are starting with the
needed initial changes. This commit adds per-arch vmlinux and removes
the hard checks on arm64, as well as disabling DWARF-based unwinding
when running in this architecture.

The flag to disable DWARF unwinding bit rot, too, so this fixes it.

Test Plan
=========

TBD

- frame pointer + dwarf unwinding should work in x86
- DWARF unwinding disabling flag should work
- we should be able to unwind in arm64 with fp

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>